+++ spinetoolbox/widgets/set_description_dialog.py
